refactor: move film component to separate file
